refactor: move permission check to main
so we can get the permission right both in tui and cli
